Pass atiData in EventTrackingContext
